requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.2.2